+++ xdump/base.py
@@ -18,6 +18,9 @@ class BaseBackend:
     schema_filename = 'dump/schema.sql'
     initial_setup_files = (schema_filename, )
     data_dir = 'dump/data/'
+    tables_sql = None
+    non_recursive_relations_query = None
+    recursive_relations_query = None
 
     # Connection
 
@@ -81,8 +84,73 @@ class BaseBackend:
         partial_tables = partial_tables or {}
         with zipfile.ZipFile(filename, 'w', zipfile.ZIP_DEFLATED) as file:
             self.write_initial_setup(file)
+            self.add_related_data(full_tables, partial_tables)
             self.write_full_tables(file, full_tables)
             self.write_partial_tables(file, partial_tables)
+
+    def add_related_data(self, full_tables, partial_tables):
+        """
+        Updates selects for partial tables to grab all objects, that are referenced by full / partial tables.
+        """
+        for table in self.tables:
+            self.update_partial_tables(table, full_tables, partial_tables)
+
+    def update_partial_tables(self, table, full_tables, partial_tables):
+        self.update_recursive_relations(table, full_tables, partial_tables)
+        self.update_non_recursive_relations(table, full_tables, partial_tables)
+
+    def update_recursive_relations(self, table, full_tables, partial_tables):
+        for foreign_key in self.get_foreign_keys(table, full_tables, recursive=True):
+            if table in partial_tables:
+                partial_tables[table] = RECURSIVE_QUERY_TEMPLATE.format(
+                    source=partial_tables[table], target=table, **foreign_key
+                )
+
+    def update_non_recursive_relations(self, table, full_tables, partial_tables):
+        for foreign_key in self.get_foreign_keys(table, full_tables):
+            sql = self.get_related_data_sql(foreign_key, full_tables, partial_tables)
+            if sql:
+                foreign_table = foreign_key['foreign_table_name']
+                if foreign_table in partial_tables:
+                    partial_tables[foreign_table] += 'UNION ' + sql
+                else:
+                    partial_tables[foreign_table] = sql
+                # Now we select more than before for given table, so we have to do check related data for it.
+                self.update_partial_tables(foreign_table, full_tables, partial_tables)
+
+    @property
+    def tables(self):
+        """
+        All non-system tables.
+        """
+        for result in self.run(self.tables_sql):
+            yield result['table_name']
+
+    def get_foreign_keys(self, table, full_tables=(), recursive=False):
+        """
+        Looks for foreign keys in the given table. Excluding ones, that will be dumped in ``full_tables``.
+        """
+        query = self.recursive_relations_query if recursive else self.non_recursive_relations_query
+        return self.run(query, {'table_name': table, 'full_tables': list(full_tables)})
+
+    def get_related_data_sql(self, foreign_key, full_tables, partial_tables):
+        """
+        Generates SQL to select related data, that is referred from another table.
+        """
+        table_name = foreign_key['table_name']
+        if table_name in full_tables:
+            source = foreign_key['table_name']
+        elif table_name in partial_tables:
+            source = '({}) T'.format(partial_tables[table_name])
+        else:
+            return
+        return '''
+            SELECT
+                *
+            FROM {foreign_table_name}
+            WHERE {foreign_column_name} IN (
+                SELECT {column_name} FROM {source}
+            )'''.format(source=source, **foreign_key)
 
     def write_initial_setup(self, file):
         self.write_schema(file)
@@ -167,3 +235,15 @@ class BaseBackend:
         Loads a data file into the database.
         """
         raise NotImplementedError
+
+
+RECURSIVE_QUERY_TEMPLATE = '''
+WITH RECURSIVE recursive_cte AS (
+  SELECT * FROM ({source}) S
+  UNION
+  SELECT T.*
+  FROM {table_name} T
+  INNER JOIN recursive_cte ON (recursive_cte.{column_name} = T.{foreign_column_name})
+)
+SELECT * FROM recursive_cte
+'''