+++ xdump/sqlite.py
@@ -1,6 +1,7 @@
 # coding: utf-8
 import sqlite3
 import subprocess
+import sys
 from csv import DictReader, DictWriter
 from io import StringIO
 from pathlib import Path
@@ -19,6 +20,7 @@ def force_string(value):
 
 
 class SQLiteBackend(BaseBackend):
+    tables_sql = "SELECT name AS table_name FROM sqlite_master WHERE type='table'"
 
     def connect(self, *args, **kwargs):
         connection = sqlite3.connect(self.dbname)
@@ -38,9 +40,30 @@ class SQLiteBackend(BaseBackend):
         cursor = self.get_cursor()
         cursor.executescript(sql)
 
-    def dump(self, *args, **kwargs):
+    def begin_immediate(self):
         cursor = self.get_cursor()
         cursor.execute('BEGIN IMMEDIATE')
+
+    def get_foreign_keys(self, table, full_tables=(), recursive=False):
+        for foreign_key in self.run('PRAGMA foreign_key_list({})'.format(table)):
+            if foreign_key['table'] in full_tables:
+                continue
+            if foreign_key['table'] == table and not recursive:
+                continue
+            if foreign_key['table'] != table and recursive:
+                continue
+            yield {
+                'foreign_table_name': foreign_key['table'],
+                'table_name': table,
+                'foreign_column_name': foreign_key['to'],
+                'column_name': foreign_key['from'],
+            }
+        if sys.version_info[:2] < (3, 6):
+            # Before 3.6 sqlite3 used to implicitly commit an open transaction in this case.
+            self.begin_immediate()
+
+    def dump(self, *args, **kwargs):
+        self.begin_immediate()
         super().dump(*args, **kwargs)
 
     def dump_schema(self):